Probability for given dice and staircase problem
We find the probability to climb 60 or fewer steps in the given no. of throws. For that, we calculated all possible cases of the dice turning up in n throws and also calculated the probability of getting the desired output by multiplying the probabilities with the no. of combinations possible. Then we subtract the value from 1 to get the probability of climbing 60 or more steps in the given no. of throws.